Fim da Integração do Código
- PRIME funcionando. (Parte desenvolvida pela Frâncila)
- Acesso à Ontologia de Colaboração. (Início da Implementação)
- Validação semi-automática para a interoperabilidade de dados entre
Serviços de Colaboração. (Início da Implementação)
- OBSERVAÇÃO: A partir deste comite, toda parte de colaboração será
desenvolvida neste projeto.

+++ src/java/br/ufjf/pgcc/plscience/dao/CoordinationServiceDAO.java
@@ -60,14 +60,14 @@ public class CoordinationServiceDAO extends GenericDAO {
         
         Query query = getEntityManager().createNativeQuery(q);
 
-        List<BigInteger> roles = query.getResultList();
+        List<Long> roles = query.getResultList();
         finish();
         if (roles != null && roles.size() > 0) {
             
             List<Long> rolesID = new ArrayList<Long>();
             
-            for(BigInteger bi : roles){
-                rolesID.add(bi.longValue());
+            for(Long bi : roles){
+                rolesID.add(bi);
             }
             
             return rolesID;
